Add files via upload
These Matlab files constitute an implementation of a 1D heat equation solver using Rothe's method and the Quasi-static method. The main file heatSolver.m allows the user to run a grid refinement study to study the order of convergence of each method at a final time, and for a chosen diffusivity value 1/eps, where eps<<1.